requirements.txt
# Python dependencies for Calendar Engine

# Google API libraries
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# iCalendar generation
icalendar>=5.0.0
pytz>=2023.3

# Configuration and utilities
PyYAML>=6.0

# Contacts-specific: Lunar calendar conversion
lunarcalendar>=3.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.0.0
